refactor(registrar): extract record action handlers to useRegistrarActions hook (Decomp 5)
Strategic Direction 1 — Decomposition (audit §8).

Extracted from RegistrarPanel.jsx (lines 1064-1163, ~100 lines) into a
dedicated custom hook. Four action handler functions moved:

- runRegistrarRecordAction: posts to /registrar/records/actions batch
  endpoint. Validates record refs (getRegistrarRecordRefs) and backend
  action availability (hasBackendAction) before calling API.
- handleStartVisit: wraps runRegistrarRecordAction for 'start_visit'.
  Shows success/error toast, reloads appointments.
- handlePayment: wraps runRegistrarRecordAction for 'mark_paid'.
  Handles partial success (some records already paid), shows toast,
  silent reload after 800ms. Was a plain function (not useCallback) —
  converted to useCallback for consistency.
- updateAppointmentStatus: maps status string to backend action
  (getRegistrarActionForStatus), calls runRegistrarRecordAction,
  reloads appointments.

NOT extracted (remain inline — simple state setters, 1-3 lines each):
- openRecordPreview: setRecordPreviewDialog({open: true, row})
- openRecordEditor: setWizardEditMode/InitialData/ShowWizard
- handleContextMenuAction: switch statement calling the above + setters

Hook API:
  useRegistrarActions({ appointments, loadAppointments })
  Returns: { runRegistrarRecordAction, handleStartVisit,
    handlePayment, updateAppointmentStatus }

Cleanup:
- Removed unused imports from RegistrarPanel.jsx:
  findRegistrarRecordBySelectionKey, hasBackendAction,
  getRegistrarActionForStatus, getRegistrarRecordRefs
  (now only used inside useRegistrarActions.js / useRegistrarData.js)
- Updated contract test: added useRegistrarActions.js to source tree

Stats:
- RegistrarPanel.jsx: 3309 → 3218 lines (−91 lines)
- useRegistrarActions.js: 130 lines (new)
- Tests: 467/467 passing (no regression)
- ESLint: 0 errors (2 pre-existing warnings remain)
- Vite build: success
